+++ rector.php
@@ -109,8 +109,9 @@ return RectorConfig::configure()
         'rex_log_entry' => Util\LogEntry::class,
         'rex_log_file' => Util\LogFile::class,
         'rex_logger' => Util\Logger::class,
-        // 'rex_markdown' => Util\Markdown::class,
+        'rex_markdown' => Util\Markdown::class,
         // 'rex_pager' => Util\Pager::class,
+        'rex_parsedown' => Util\Parsedown::class,
         'rex_path' => Filesystem\Path::class,
         'rex_path_default_provider' => Filesystem\DefaultPathProvider::class,
         'rex_sql' => Database\Sql::class,
