+++ redaxo/src/addons/install/pages/index.php
@@ -1,5 +1,7 @@
 <?php
 
+use Redaxo\Core\Util\Markdown;
+
 $addon = rex_addon::get('install');
 
 echo rex_view::title($addon->i18n('title'));
@@ -10,7 +12,7 @@ if ('reload' === rex_request('func', 'string')) {
 
 $markdown = static function (string $content): string {
     $fragment = new rex_fragment();
-    $fragment->setVar('content', rex_markdown::factory()->parse($content), false);
+    $fragment->setVar('content', Markdown::factory()->parse($content), false);
 
     return $fragment->parse('core/page/readme.php');
 };
