+++ redaxo/src/core/pages/packages.changelog.php
@@ -2,15 +2,16 @@
 
 use Redaxo\Core\Filesystem\File;
 use Redaxo\Core\Translation\I18n;
+use Redaxo\Core\Util\Markdown;
 
 $content = '';
 
 $package = rex_addon::get(rex_request('package', 'string'));
 
 if (is_readable($package->getPath('CHANGELOG.md'))) {
-    [$readmeToc, $readmeContent] = rex_markdown::factory()->parseWithToc(File::require($package->getPath('CHANGELOG.md')), 1, 2, [
-        rex_markdown::SOFT_LINE_BREAKS => false,
-        rex_markdown::HIGHLIGHT_PHP => true,
+    [$readmeToc, $readmeContent] = Markdown::factory()->parseWithToc(File::require($package->getPath('CHANGELOG.md')), 1, 2, [
+        Markdown::SOFT_LINE_BREAKS => false,
+        Markdown::HIGHLIGHT_PHP => true,
     ]);
     $fragment = new rex_fragment();
     $fragment->setVar('content', $readmeContent, false);
