+++ redaxo/src/core/pages/packages.help.php
@@ -2,6 +2,7 @@
 
 use Redaxo\Core\Filesystem\File;
 use Redaxo\Core\Translation\I18n;
+use Redaxo\Core\Util\Markdown;
 
 $content = '';
 
@@ -18,18 +19,18 @@ if (is_readable($package->getPath('help.php'))) {
     $package->includeFile('help.php');
     $content .= ob_get_clean();
 } elseif (is_readable($package->getPath('README.' . I18n::getLanguage() . '.md'))) {
-    [$readmeToc, $readmeContent] = rex_markdown::factory()->parseWithToc(File::require($package->getPath('README.' . I18n::getLanguage() . '.md')), 2, 3, [
-        rex_markdown::SOFT_LINE_BREAKS => false,
-        rex_markdown::HIGHLIGHT_PHP => true,
+    [$readmeToc, $readmeContent] = Markdown::factory()->parseWithToc(File::require($package->getPath('README.' . I18n::getLanguage() . '.md')), 2, 3, [
+        Markdown::SOFT_LINE_BREAKS => false,
+        Markdown::HIGHLIGHT_PHP => true,
     ]);
     $fragment = new rex_fragment();
     $fragment->setVar('content', $readmeContent, false);
     $fragment->setVar('toc', $readmeToc, false);
     $content .= $fragment->parse('core/page/docs.php');
 } elseif (is_readable($package->getPath('README.md'))) {
-    [$readmeToc, $readmeContent] = rex_markdown::factory()->parseWithToc(File::require($package->getPath('README.md')), 2, 3, [
-        rex_markdown::SOFT_LINE_BREAKS => false,
-        rex_markdown::HIGHLIGHT_PHP => true,
+    [$readmeToc, $readmeContent] = Markdown::factory()->parseWithToc(File::require($package->getPath('README.md')), 2, 3, [
+        Markdown::SOFT_LINE_BREAKS => false,
+        Markdown::HIGHLIGHT_PHP => true,
     ]);
     $fragment = new rex_fragment();
     $fragment->setVar('content', $readmeContent, false);
