+++ src/Util/Parsedown.php
@@ -1,112 +1,16 @@
 <?php
 
-/**
- * Markdown parser.
- */
-class rex_markdown
-{
-    use rex_factory_trait;
+namespace Redaxo\Core\Util;
 
-    public const SOFT_LINE_BREAKS = 'soft_line_breaks';
-    public const HIGHLIGHT_PHP = 'highlight_php';
+use ParsedownExtra;
+use rex_string;
 
-    private function __construct() {}
-
-    /**
-     * @return static
-     */
-    public static function factory()
-    {
-        $class = static::getFactoryClass();
-        return new $class();
-    }
-
-    /**
-     * Parses markdown code.
-     *
-     * @param string $code Markdown code
-     * @param array<self::*, bool> $options
-     *
-     * @return string HTML code
-     */
-    public function parse($code, $options = [])
-    {
-        $parser = new rex_parsedown();
-        $parser->setBreaksEnabled($options[self::SOFT_LINE_BREAKS] ?? true);
-        $parser->highlightPhp = $options[self::HIGHLIGHT_PHP] ?? false;
-
-        return rex_string::sanitizeHtml($parser->text($code));
-    }
-
-    /**
-     * Parses markdown code and extracts a table-of-content.
-     *
-     * @param string $code Markdown code
-     * @param int $topLevel Top included headline level for TOC, e.g. `1` for `<h1>`
-     * @param int $bottomLevel Bottom included headline level for TOC, e.g. `6` for `<h6>`
-     * @param array<self::*, bool> $options
-     *
-     * @return list{string, string} tupel of table-of-content and content
-     */
-    public function parseWithToc($code, $topLevel = 2, $bottomLevel = 3, $options = [])
-    {
-        $parser = new rex_parsedown();
-        $parser->setBreaksEnabled($options[self::SOFT_LINE_BREAKS] ?? true);
-        $parser->highlightPhp = $options[self::HIGHLIGHT_PHP] ?? false;
-
-        $parser->generateToc = true;
-        $parser->topLevel = $topLevel;
-        $parser->bottomLevel = $bottomLevel;
-
-        $content = rex_string::sanitizeHtml($parser->text($code));
-        $headers = $parser->headers;
-
-        $previous = $topLevel - 1;
-        $toc = '';
-
-        foreach ($headers as $header) {
-            $level = $header['level'];
-
-            if ($level > $previous) {
-                if ($level > $previous + 1) {
-                    $message = 'The headline structure in the given markdown document is malformed, ';
-                    if ($previous < $topLevel) {
-                        $message .= "it starts with a h$level instead of a h$topLevel.";
-                    } else {
-                        $message .= "a h$previous is followed by a h$level, but only a h" . ($previous + 1) . ' or lower is allowed.';
-                    }
-
-                    throw new rex_exception($message);
-                }
-
-                $toc .= "<ul>\n";
-                $previous = $level;
-            } elseif ($level < $previous) {
-                for (; $level < $previous; --$previous) {
-                    $toc .= "</li>\n";
-                    $toc .= "</ul>\n";
-                }
-            } else {
-                $toc .= "</li>\n";
-            }
-
-            $toc .= "<li>\n";
-            $toc .= '<a href="#' . rex_escape($header['id']) . '">' . rex_escape($header['text']) . "</a>\n";
-        }
-
-        for (; $previous > $topLevel - 1; --$previous) {
-            $toc .= "</li>\n";
-            $toc .= "</ul>\n";
-        }
-
-        return [$toc, $content];
-    }
-}
+use const E_DEPRECATED;
 
 /**
  * @internal
  */
-final class rex_parsedown extends ParsedownExtra
+final class Parsedown extends ParsedownExtra
 {
     /** @var bool */
     public $highlightPhp = false;
